And now the iOS code compiles as well.

+++ src/ios/Calendar.m
@@ -124,7 +124,7 @@
     myEvent.calendar = calendar;
     
     // added by Vt
-    EKAlarm *reminder = [EKAlarm alarmWithRelativeOffset:-1*alarmTime*60];
+    EKAlarm *reminder = [EKAlarm alarmWithRelativeOffset:-1*alarmTime.intValue*60];
     [myEvent addAlarm:reminder];
     
     NSError *error = nil;
